Build the os.bin
This is a big patch that should probably be broken down into a series.

Build the os.bin appending the kernel to the booltloader and add as much
zero padding as the expected kernel size.

Factor out the kernel code from boot.asm and give it a new home, kernel.asm,
which is compiled to assemble the kernel.bin.

Add the ata_lba_read fucntion to boot.asm since the BIOS int 0x13 can't
be accessed in protected mode.

Signed-off-by: Claudio Carvalho <claudiodecarvalho@gmail.com>